Create dazzle_import.php
If you use Endicia DAZzle to print shipping labels for USPS packages you can use this to mark orders as Completed in bulk instead of doing them one at a time.  Go to Postage Log and export the orders you want to mark as completed as a .txt file, and then upload the file with this to bulk mark them completed.